+++ CHANGELOG.md
@@ -10,9 +10,10 @@
 <Viewer initialRotation={90} />
 ```
 
-**Bug fix**
+**Bug fixes**
 
 -   Can't render certain PDF documents that contain an annotation whose destination consists of non alphabetical characters
+-   The `startPageIndex` and `endPageIndex` properties of `SelectionData` aren't correct
 
 ## v3.7.0